+++ src/org/wordpress/android/ui/AuthenticatedWebViewActivity.java
@@ -1,20 +1,13 @@
 
 package org.wordpress.android.ui;
 
-import java.io.UnsupportedEncodingException;
-import java.lang.reflect.Type;
-import java.net.URLEncoder;
-import java.util.Map;
-
 import android.content.Context;
 import android.content.Intent;
 import android.net.Uri;
 import android.os.Bundle;
-import android.webkit.HttpAuthHandler;
 import android.webkit.WebChromeClient;
 import android.webkit.WebSettings;
 import android.webkit.WebView;
-import android.webkit.WebViewClient;
 import android.widget.Toast;
 
 import com.actionbarsherlock.view.Menu;
@@ -25,8 +18,14 @@ import com.google.gson.reflect.TypeToken;
 
 import org.wordpress.android.R;
 import org.wordpress.android.WordPress;
-import org.wordpress.passcodelock.AppLockManager;
 import org.wordpress.android.models.Blog;
+import org.wordpress.android.util.WPWebViewClient;
+import org.wordpress.passcodelock.AppLockManager;
+
+import java.io.UnsupportedEncodingException;
+import java.lang.reflect.Type;
+import java.net.URLEncoder;
+import java.util.Map;
 
 /**
  * Activity for displaying WordPress content in a webview which may require authentication.
@@ -52,7 +51,7 @@ public class AuthenticatedWebViewActivity extends WebViewActivity {
             finish();
         }
 
-        mWebView.setWebViewClient(new WordPressWebViewClient(mBlog));
+        mWebView.setWebViewClient(new WPWebViewClient(mBlog));
 
         mWebView.getSettings().setCacheMode(WebSettings.LOAD_NO_CACHE);
         mWebView.getSettings().setSavePassword(false);
@@ -66,7 +65,7 @@ public class AuthenticatedWebViewActivity extends WebViewActivity {
 
     /**
      * Get the URL of the WordPress login page.
-     * 
+     *
      * @return URL of the login page.
      */
     protected String getLoginUrl() {
@@ -88,13 +87,13 @@ public class AuthenticatedWebViewActivity extends WebViewActivity {
                 return mBlog.getUrl().replace("xmlrpc.php", "wp-login.php");
             }
         }
-        
+
         return loginURL;
     }
 
     /**
      * Login to the WordPress blog and load the specified URL.
-     * 
+     *
      * @param url URL to be loaded in the webview.
      */
     protected void loadAuthenticatedUrl(String url) {
@@ -105,40 +104,6 @@ public class AuthenticatedWebViewActivity extends WebViewActivity {
             mWebView.postUrl(getLoginUrl(), postData.getBytes());
         } catch (UnsupportedEncodingException e) {
             e.printStackTrace();
-        }
-    }
-
-    /**
-     * WebViewClient that is capable of handling HTTP authentication requests using the HTTP
-     * username and password of the blog configured for this activity.
-     */
-    private class WordPressWebViewClient extends WebViewClient {
-        private Blog blog;
-
-        WordPressWebViewClient(Blog blog) {
-            super();
-            this.blog = blog;
-        }
-
-        @Override
-        public boolean shouldOverrideUrlLoading(WebView view, String url) {
-            // Found a bug on some pages where there is an incorrect
-            // auto-redirect to file:///android_asset/webkit/.
-            if ( !url.equals("file:///android_asset/webkit/") ) {
-                view.loadUrl(url);
-            }
-            return true;
-        }
-
-        @Override
-        public void onPageFinished(WebView view, String url) {
-            
-        }
-
-        @Override
-        public void onReceivedHttpAuthRequest(WebView view,
-                HttpAuthHandler handler, String host, String realm) {
-            handler.proceed(blog.getHttpuser(), blog.getHttppassword());
         }
     }
 
@@ -161,9 +126,9 @@ public class AuthenticatedWebViewActivity extends WebViewActivity {
             if (progress == 100) {
                 setTitle(webView.getTitle());
             }
-        } 
+        }
     }
-    
+
     @Override
     public boolean onCreateOptionsMenu(Menu menu) {
         super.onCreateOptionsMenu(menu);
@@ -176,7 +141,7 @@ public class AuthenticatedWebViewActivity extends WebViewActivity {
     public boolean onOptionsItemSelected(final MenuItem item) {
         if (mWebView == null)
             return false;
-        
+
         int itemID = item.getItemId();
         if (itemID == R.id.menu_refresh) {
             mWebView.reload();
