+++ src/org/wordpress/android/ui/themes/ThemePreviewFragment.java
@@ -6,11 +6,9 @@ import android.support.v4.app.Fragment;
 import android.view.LayoutInflater;
 import android.view.View;
 import android.view.ViewGroup;
-import android.webkit.HttpAuthHandler;
 import android.webkit.WebChromeClient;
 import android.webkit.WebSettings;
 import android.webkit.WebView;
-import android.webkit.WebViewClient;
 import android.widget.ProgressBar;
 
 import com.actionbarsherlock.app.SherlockFragment;
@@ -21,6 +19,7 @@ import com.actionbarsherlock.view.MenuItem;
 import org.wordpress.android.R;
 import org.wordpress.android.WordPress;
 import org.wordpress.android.models.Blog;
+import org.wordpress.android.util.WPWebViewClient;
 
 import java.io.UnsupportedEncodingException;
 import java.net.URLEncoder;
@@ -34,53 +33,53 @@ public class ThemePreviewFragment extends SherlockFragment {
     public static final String TAG = ThemePreviewFragment.class.getName();
     private static final String ARGS_THEME_ID = "theme_id";
     private static final String ARGS_PREVIEW_URL = "preview_url";
-    
-    // sample desktop user-agent to force desktop view of site 
+
+    // sample desktop user-agent to force desktop view of site
     private static final String DESKTOP_UA = "Mozilla/5.0 (X11; U; Linux i686; en-US; rv:1.9.0.4) Gecko/20100101 Firefox/4.0";
 
-    
+
     private ThemePreviewFragmentCallback mCallback;
     private WebView mWebView;
     private Blog mBlog;
     private ProgressBar mProgressBar;
     private String mThemeId;
     private String mPreviewURL;
-    
+
     public interface ThemePreviewFragmentCallback {
         public void onResume(Fragment fragment);
         public void onPause(Fragment fragment);
         public void onActivateThemeClicked(String themeId, Fragment fragment);
     }
-    
-    
+
+
     public static ThemePreviewFragment newInstance(String themeId, String previewURL) {
         ThemePreviewFragment fragment = new ThemePreviewFragment();
-        
+
         Bundle args = new Bundle();
         args.putString(ARGS_THEME_ID, themeId);
         args.putString(ARGS_PREVIEW_URL, previewURL);
         fragment.setArguments(args);
-        
+
         return fragment;
     }
-    
+
     @Override
     public void onCreate(Bundle savedInstanceState) {
         super.onCreate(savedInstanceState);
         setHasOptionsMenu(true);
     }
-    
+
     @Override
     public void onAttach(Activity activity) {
         super.onAttach(activity);
-        
+
         try {
             mCallback = (ThemePreviewFragmentCallback) activity;
         } catch (ClassCastException e) {
             throw new ClassCastException(activity.toString() + " must implement ThemePreviewFragmentCallback");
         }
     }
-    
+
     @Override
     public void onResume() {
         super.onResume();
@@ -88,24 +87,24 @@ public class ThemePreviewFragment extends SherlockFragment {
         setHasOptionsMenu(true);
         setMenuVisibility(true);
     }
-    
+
     @Override
     public void onPause() {
         super.onPause();
         mCallback.onPause(this);
     }
-    
+
     public String getThemeId() {
         if (mThemeId != null) {
             return mThemeId;
         } else if (getArguments() != null) {
-            mThemeId = getArguments().getString(ARGS_THEME_ID); 
+            mThemeId = getArguments().getString(ARGS_THEME_ID);
             return mThemeId;
         } else {
             return null;
         }
     }
-    
+
     private String getPreviewURL() {
         if (mPreviewURL != null) {
             return mPreviewURL;
@@ -113,10 +112,10 @@ public class ThemePreviewFragment extends SherlockFragment {
             mPreviewURL = getArguments().getString(ARGS_PREVIEW_URL);
             return mPreviewURL;
         } else {
-            return null; 
+            return null;
         }
     }
-    
+
     public void load(String themeId, String previewURL) {
         mThemeId = themeId;
         mPreviewURL = previewURL;
@@ -128,18 +127,18 @@ public class ThemePreviewFragment extends SherlockFragment {
 
         mBlog = WordPress.getCurrentBlog();
         String previewURL = getPreviewURL();
-        
+
         if (previewURL == null || mBlog == null)
             getActivity().getSupportFragmentManager().beginTransaction().remove(this).commit();
-        
-        View view = inflater.inflate(R.layout.webview, container, false);        
-        
+
+        View view = inflater.inflate(R.layout.webview, container, false);
+
         mWebView = (WebView) view.findViewById(R.id.webView);
         mProgressBar = (ProgressBar) view.findViewById(R.id.progressBar);
         mWebView.getSettings().setUserAgentString(DESKTOP_UA);
         mWebView.getSettings().setBuiltInZoomControls(true);
         mWebView.setScrollBarStyle(View.SCROLLBARS_INSIDE_OVERLAY);
-        
+
         mWebView.setWebChromeClient(new WebChromeClient() {
             @Override
             public void onProgressChanged(WebView view, int newProgress) {
@@ -148,28 +147,28 @@ public class ThemePreviewFragment extends SherlockFragment {
                     mProgressBar.setVisibility(ProgressBar.VISIBLE);
                 }
                 mProgressBar.setProgress(newProgress);
-                
+
                 if(newProgress == 100) {
                     mProgressBar.setVisibility(ProgressBar.GONE);
                 }
             }
         });
-        
+
         mWebView.setWebViewClient(new WordPressWebViewClient(mBlog));
 
         mWebView.getSettings().setCacheMode(WebSettings.LOAD_NO_CACHE);
         mWebView.getSettings().setSavePassword(false);
-        
+
         loadAuthenticatedUrl(previewURL);
-        
+
         return view;
     }
 
     private void refreshViews() {
-        loadAuthenticatedUrl(getPreviewURL());   
+        loadAuthenticatedUrl(getPreviewURL());
     }
 
-    
+
     /**
      * Login to the WordPress blog and load the specified URL.
      *
@@ -187,17 +186,14 @@ public class ThemePreviewFragment extends SherlockFragment {
             e.printStackTrace();
         }
     }
-    
+
     /**
      * WebViewClient that is capable of handling HTTP authentication requests using the HTTP
      * username and password of the blog configured for this activity.
      */
-    private class WordPressWebViewClient extends WebViewClient {
-        private Blog blog;
-
+    private class WordPressWebViewClient extends WPWebViewClient {
         WordPressWebViewClient(Blog blog) {
-            super();
-            this.blog = blog;
+            super(blog);
         }
 
         @Override
@@ -210,24 +206,13 @@ public class ThemePreviewFragment extends SherlockFragment {
         public void onPageFinished(WebView view, String url) {
             mProgressBar.setVisibility(ProgressBar.GONE);
         }
-
-        @Override
-        public void onReceivedHttpAuthRequest(WebView view, HttpAuthHandler handler, String host, String realm) {
-            handler.proceed(blog.getHttpuser(), blog.getHttppassword());
-        }
-
-        @Override
-        public void onReceivedError(WebView view, int errorCode, String description, String failingUrl) {
-            super.onReceivedError(view, errorCode, description, failingUrl);
-        }
-
     }
-    
+
     @Override
     public void onCreateOptionsMenu(Menu menu, MenuInflater inflater) {
         inflater.inflate(R.menu.theme_preview, menu);
     }
-    
+
     @Override
     public boolean onOptionsItemSelected(MenuItem item) {
         switch (item.getItemId()) {
@@ -238,7 +223,7 @@ public class ThemePreviewFragment extends SherlockFragment {
                 return super.onOptionsItemSelected(item);
         }
     }
-    
+
     @Override
     public void onPrepareOptionsMenu(Menu menu) {
         menu.removeItem(R.id.menu_search);
